text-extractor-function/requirements.txt
functions-framework==3.*
google-cloud-storage==3.3.0
google-cloud-documentai==3.5.0
google-cloud-pubsub==2.31.1
pandas==2.3.2
openpyxl==3.1.5
python-docx==1.2.0